+++ WordPress/src/main/java/org/wordpress/android/ui/suggestion/util/SuggestionServiceConnectionManager.java
@@ -9,13 +9,12 @@ import android.os.IBinder;
 import org.wordpress.android.ui.suggestion.service.SuggestionService;
 
 public class SuggestionServiceConnectionManager implements ServiceConnection {
-
     private final Context mContext;
-    private final int mRemoteBlogId;
+    private final long mRemoteBlogId;
     private boolean mAttemptingToBind = false;
     private boolean mBindCalled = false;
 
-    public SuggestionServiceConnectionManager(Context context, int remoteBlogId) {
+    public SuggestionServiceConnectionManager(Context context, long remoteBlogId) {
         mContext = context;
         mRemoteBlogId = remoteBlogId;
     }
